refactor: code
- extract outbound functions
- preparing to implement udp assosiate
