+++ src/pgs_defs.h
@@ -4,6 +4,7 @@
 #include <string.h>
 
 #define BUFSIZE_16K 16 * 1024
+#define BUFSIZE_4K 4 * 1024
 #define memzero(buf, n) (void)memset(buf, 0, n)
 
 #endif
